requirements.txt
aiogram==3.15.0
aiohttp==3.9.5
requests==2.32.3
python-dotenv==1.0.1